feat: 사용자 추방 api #242
